move CorrectedOutputFunctional to constructions and define a project rule
